requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pymongo==4.9.0
pydantic==2.9.2
pydantic-settings==2.6.1
PyJWT==2.9.0
cryptography==43.0.3
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
python-dotenv==1.0.1
requests==2.32.3
pyotp==2.9.0
